requirements.txt
streamlit==1.31.1
# chromadb==0.3.29
joblib==1.4.0
langchain==0.1.16
langchain_community==0.0.32
langchain_core==0.1.52
langchain_openai==0.1.3
# langchainhub
lime==0.2.0.1
Pillow==9.5.0
# Pillow==10.2.0
plotly==5.15.0
wordcloud==1.9.3
matplotlib==3.8.4 #3.7.1
nltk==3.8.1
numpy==1.23.4
statsmodels==0.14.0 
faiss-cpu==1.8.0
tensorflow==2.12.0
pandas==2.2.2
